Get some of the builder_test.go tests working again.
TestBuilderPrepare_ISOChecksumType is removed and changed to verify that
there is a deprecation notice. This notice automatically tells the user
that `packer fix` can fix the problem.

TestBuilderPrepare_ISOChecksum now iterates through all the valid
checksum type strings as well as permitting none; this is consistent
with permitted packer types.

TestBuilderPrepare_ISOUrl is still failing at this point.