Fixes for the iso builder tests.
We no longer check if ISOUrls is being altered, the current packer sdk
appears to be smart enough to deal either or.

Add in a test to make sure that ISOName takes precedence over ISO_URL.
This mimics the previous behavior but IMO it probably should be an
either or thing with an error message if you set both.